requirements.txt
fastapi
uvicorn[standard]
httpx
PyYAML
pydantic
openai
colorlog